agent: Move vfio device code to vfio_device_handler
Move vfio device code to vfio_device_handler to simplify the code.
